Team1/keh

+++ src/main/java/com/meomureum/springboot/dto/FileuploadDTO.java
@@ -9,6 +9,6 @@ public class FileuploadDTO {
 	private String target_code;
 	private String file_path;
 	private String file_name;
-	private int file_size;
-	private int file_order;
+	private long file_size; // long으로 바꿈
+	private int file_order; 
 }
